feat/#85: 일별 평균 혈당 조회 api 개발

+++ src/main/java/com/epi/epilog/app/dto/LogsResponseDto.java
@@ -64,9 +64,9 @@ public class LogsResponseDto {
     @NoArgsConstructor(access =  AccessLevel.PRIVATE)
     public static class DayAvgBloodSugar {
         private String date;
-        private Long average;
-        private Long preAverage;
-        private Long postAverage;
+        private Double average;
+        private Double preAverage;
+        private Double postAverage;
     }
 
     /**
